removed prod test cases
