+++ e2e/cypress/integration/enterprise/system_console/compliance/s3_bucket_storage_spec.js
@@ -7,7 +7,6 @@
 // - Use element ID when selecting an element. Create one if none.
 // ***************************************************************
 
-// Stage: @prod
 // Group: @enterprise @system_console @compliance_export @not_cloud
 
 // Requires "mattermost-minio" docker instance to be accessible at http://localhost:9000
