+++ e2e/cypress/integration/files_and_attachments/image_link_preview_spec.js
@@ -96,41 +96,6 @@ describe('Image Link Preview', () => {
         cy.findAllByLabelText('file thumbnail').should('be.visible').and('have.length', 4);
     });
 
-    it('MM-T332 Image link preview - Bitly links for images and YouTube -- KNOWN ISSUE: MM-40448', () => {
-        // # Youtube link and image link
-        const links = ['https://bit.ly/2NlYsOr', 'https://bit.ly/2wqEbjw'];
-
-        links.forEach((link) => {
-            // # Post a link to an externally hosted image
-            cy.postMessage(link);
-
-            cy.getLastPostId().then((postId) => {
-                // * Verify it renders correctly on center
-                cy.get(`#post_${postId}`).should('be.visible').within(() => {
-                    cy.findByLabelText('Toggle Embed Visibility').
-                        should('be.visible').and('have.attr', 'data-expanded', 'true');
-                    cy.findByLabelText('file thumbnail').should('be.visible');
-                });
-
-                // # Click the collapse arrows to collapse the image preview
-                cy.get(`#post_${postId}`).findByLabelText('Toggle Embed Visibility').
-                    click().
-                    should('have.attr', 'data-expanded', 'false');
-
-                // * Observe it collapses
-                cy.get(`#post_${postId}`).findByLabelText('file thumbnail').should('not.exist');
-
-                // # Click the expand arrows to expand the image preview again
-                cy.get(`#post_${postId}`).findByLabelText('Toggle Embed Visibility').
-                    click().
-                    should('have.attr', 'data-expanded', 'true');
-
-                // * Observe it expand
-                cy.get(`#post_${postId}`).findByLabelText('file thumbnail').should('be.visible');
-            });
-        });
-    });
-
     it('MM-T2389 Inline markdown image links open with preview modal', () => {
         // Go to home channel
         cy.visit(offTopicUrl);
